requirements.txt
aiogram>=3.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp-socks>=0.8.0
openpyxl>=3.1.5
